feat: support plugins and `config.wpRelease` options in package.json
This enables this module to be used for a wider variety of plugins and
themes. Now any project using this can easily add a custom ignore list
and entry file to update the version number in that WordPress is
concerned with. Additionally, the zip functionality has been moved into
node rather than spawning a subprocess to run `zip`.